Improve performance by using simpler delays
Before this commit fdelaylti was used which was accurate to the microsecond, but required more CPU power to process due to the 2nd order allpass delays.

Listening to a slowly moving contact proved that fdelay (an interpolated delay function) is indistinguishable from this effect, while significantly improving on the processing requirements.

Using a simple delay in the contact rotator proved insufficient, as lobes would audibly jump leading to an unwanted effect.

This same effect however is necessary in the compensator. That circuit used discrete delays and wasn't smooth.